Refactored new classes into namespaces
