+++ src/main/java/com/ddoerr/scriptit/bus/KeyBindingBusExtension.java
@@ -1,5 +1,6 @@
-package com.ddoerr.scriptit;
+package com.ddoerr.scriptit.bus;
 
+import com.ddoerr.scriptit.ScriptItMod;
 import com.ddoerr.scriptit.api.util.KeyBindingHelper;
 import com.ddoerr.scriptit.dependencies.Resolver;
 import net.minecraft.client.options.KeyBinding;
@@ -8,7 +9,6 @@ import net.minecraft.util.Identifier;
 import net.minecraft.util.Tickable;
 
 import java.util.*;
-import java.util.function.Consumer;
 
 public class KeyBindingBusExtension implements Tickable{
     EventBus eventBus;
