+++ src/main/java/com/ddoerr/scriptit/loader/EventImpl.java
@@ -1,6 +1,6 @@
 package com.ddoerr.scriptit.loader;
 
-import com.ddoerr.scriptit.EventBus;
+import com.ddoerr.scriptit.bus.EventBus;
 import com.ddoerr.scriptit.api.events.Event;
 import com.ddoerr.scriptit.api.libraries.NamespaceRegistry;
 import com.ddoerr.scriptit.dependencies.Resolver;
