+++ src/main/java/com/ddoerr/scriptit/scripts/LifeCycle.java
@@ -1,4 +1,4 @@
-package com.ddoerr.scriptit;
+package com.ddoerr.scriptit.scripts;
 
 public enum LifeCycle {
     Instant,
